src/main/java/org/ieeervce/api/siterearnouveau/controller/EventsController.java
package org.ieeervce.api.siterearnouveau.controller;

import org.ieeervce.api.siterearnouveau.dto.ResultsDTO;
import org.ieeervce.api.siterearnouveau.entity.Event;
import org.ieeervce.api.siterearnouveau.service.EventsService;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/event")
public class EventsController {
    private final EventsService eventsService;

    public EventsController(EventsService eventsService) {
        this.eventsService = eventsService;
    }

    @GetMapping
    public ResultsDTO<Page<Event>> list(Pageable pageable) {
        Page<Event> events = eventsService.list(pageable);
        return new ResultsDTO<>(events);
    }
}
